requirements.txt
requests==2.31.0
aiohttp==3.9.1
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
apscheduler==3.10.4
pydantic==2.5.0
alibabacloud-oss2==2.18.4
alibabacloud-dysmsapi20170525==3.0.6
